split the command field for stdin input
Modern crontab(5) supports a little known feature of being able to feed
the command to run input on stdin.  From the man page:

    The  "sixth"  field (the rest of the line) specifies the command to
    be run.  The entire command portion of the line, up to a newline or
    % character, will be executed by /bin/sh or by the shell specified
    in the SHELL variable of the cronfile.  Percent-signs (%) in the
    command, unless escaped with  backslash (\), will be changed into
    newline characters, and all data after the first % will be sent to
    the command as standard input.

Adds a `Stdin` field to the entry struct, which contains a single string
(which may be multiple lines if there were multiple %'s given).

The tests for this includes a use-case I use pretty often, specifically
redirecting output to a dated file.

What's not supported (or tested) is what happens when the command starts
with a %, which is apparently undefined? (or feeds the entire command as
stdin to the shell?).